Move Trailers iterator to "block" module
Extraction of `Trailer`s from text is performed in the blocks module.
Hence the `Trailers` iterator was based on the `Blocks` iterator. It
makes therefore sense to move the `Trailers` iterator over to `block`.
This also results in `trailer` not depending on `block` any more.
