+++ lib/src/message/block.rs
@@ -16,7 +16,8 @@
 //! well as an iterator for extracting the blocks from a sequence of lines.
 //!
 
-use message::trailer::Trailer;
+use message::trailer::{self, Trailer};
+use std::collections::VecDeque;
 use std::str::FromStr;
 
 
@@ -118,6 +119,65 @@ impl<I, S> Iterator for Blocks<I, S>
 }
 
 
+/// Iterator extracting trailers from a sequence of strings representing lines
+///
+/// This iterator extracts all trailers from a text provided by the wrapped
+/// iterator over the text's lines. Blocks of lines which contain regular lines
+/// of text are ignored. Only trailers which are part of a pure block of
+/// trailers, delimited by blank lines, are returned by the iterator.
+///
+pub struct Trailers<I, S>
+    where I: Iterator<Item = S>,
+          S: AsRef<str>
+{
+    blocks: Blocks<I, S>,
+    buf: VecDeque<Trailer>,
+}
+
+impl<I, S> Trailers<I, S>
+    where I: Iterator<Item = S>,
+          S: AsRef<str>
+{
+    pub fn only_dit(self) -> trailer::DitTrailers<Self> {
+        self.into()
+    }
+}
+
+impl<I, S> From<I> for Trailers<I, S>
+    where I: Iterator<Item = S>,
+          S: AsRef<str>
+{
+    fn from(lines: I) -> Self {
+        Trailers {
+            blocks: Blocks::from(lines),
+            buf: VecDeque::new(),
+        }
+    }
+}
+
+impl<I, S> Iterator for Trailers<I, S>
+    where I: Iterator<Item = S>,
+          S: AsRef<str>
+{
+    type Item = Trailer;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        loop {
+            if let Some(trailer) = self.buf.pop_front() {
+                return Some(trailer);
+            }
+
+            match self.blocks.next() {
+                Some(Block::Trailer(trailers)) => self.buf = VecDeque::from(trailers),
+                None => return None,
+                _ => {},
+            }
+        }
+    }
+
+}
+
+
 
 
 #[cfg(test)]
@@ -211,5 +271,50 @@ mod tests {
         }
 
         assert!(!blocks.next().is_some())
+    }
+
+    // Trailers tests
+
+    #[test]
+    fn trailers_iter() {
+        let mut trailers = Trailers::from(vec![
+            "Foo-bar: bar",
+            "",
+            "Space: the final frontier.",
+            "These are the voyages...",
+            "",
+            "And then he",
+            "said: engage!",
+            "",
+            "",
+            "Signed-off-by: Spock",
+            "Dit-status: closed",
+            "Multi-line-trailer: multi",
+            "  line",
+            "  content"
+        ].into_iter());
+
+        {
+            let (key, _) = trailers.next().expect("Failed to parse trailer1").into();
+            assert_eq!(key, "Foo-bar".to_string().into());
+        }
+
+        {
+            let (key, _) = trailers.next().expect("Failed to parse trailer2").into();
+            assert_eq!(key, "Signed-off-by".to_string().into());
+        }
+
+        {
+            let (key, _) = trailers.next().expect("Failed to parse trailer3").into();
+            assert_eq!(key, "Dit-status".to_string().into());
+        }
+
+        {
+            let (key, value) = trailers.next().expect("Failed to parse trailer4").into();
+            assert_eq!(key, "Multi-line-trailer".to_string().into());
+            assert_eq!(value, TrailerValue::String("multi  line  content".to_string()));
+        }
+
+        assert!(!trailers.next().is_some())
     }
 }
